requirements.txt
langchain
langchain-google-genai
psycopg2-binary
SQLAlchemy
python-dotenv